Chnged int to Long in the Loan entity. Package constants moved to enums.

+++ library-system-persistence/src/main/java/cz/muni/fi/pa165/library/persistence/entity/Loan.java
@@ -17,13 +17,13 @@ import javax.persistence.Temporal;
 public class Loan {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
